+++ src/universal/Entity/Task/Task.php
@@ -153,6 +153,18 @@ class Task implements ProxyInterface, TaskInterface, AutomatedInterface
     }
 
     /**
+     * To override task's uuid
+     * @param string $id
+     * @return Task
+     */
+    public function setId(string $id): Task
+    {
+        $this->id = $id;
+
+        return $this;
+    }
+
+    /**
      * {@inheritdoc}
      */
     public function setCode(CodeInterface $code): TaskInterface